Deleted the FrameLayout for a ViewPager
Deleted all the methods who configure the fragments.

Set the viewPager on MainActivity to handle the click of the user on the different tabs

Moved the Viewpager and the TabLayout from the ViewPager Layout to the Main activity Layout.

Setup the TabLayout to show and manipulate tabs on the main sreen